Move opt-in logic to separate file
